Add BLE scanning in Central role
Scan result from BLE scanning is added to the arraylist backing the
array adapter that is set to the FlingSwipeAdapterView.

Add defensive checks against exceptions, some of which seen in testing
such as:
 NPE in runOnUiThread runnable in onScanResult() callback after BLE scan
 IndexOutOfBoundsException in fling listener's removeFirstObjectInAdapter()

Add uses-feature and uses-permission in manifest file.
Update butterknife library version to current.

add defensive checks in listener and callback for exceptions