Add BLE advertise functionality
Analogous to approach in scan for BLE devices method, we have one method to start
and stop BLE advertise, and create just a local BluetoothLeAdvertiser object in
the method instead of declaring a class field.

We generated a custom service uuid for the advertise data.

We start advertise in onResume(), and temporarily comment out code to start
scan since device cannot do both at same time.

Minor refactoring of code to scan for BLE device, mostly to conform with code to
advertise, including adding a Toast message in scan-failed callback similar to
Toast message in advertise-startfailure callback.