Add repeating alarm to start scan so device stays awake to advertise/scan
First, we change logic in code to start scan to first stop advertising.
Similarly, we change logic in code to stop scan so that we start advertising after
scanning has stopped.

Remove timeout setting when starting advertise because advertising would be stopped
by the periodic method call to start the scan.

Add a broadcast receiver class to respond to repeating alarm; need to inject in the
activity object. That is so that we can call the method to scan for recognizable ble
devices, but only if device is ready for interaction with user e.g. screen in on,
and (likely anyway) is not in power save mode. If not calling method to scan for
ble devices, we defensively stop/start advertise at each repeating alarm, just to be
sure the device is advertising in background.

Add/improve comments.