Add code to perform Central (client) role
On touching a swipe card, want to connect (in Central aka client role) to the
remote device (in Peripheral aka server role).

Refactor the array adapter to have parameter type Pair<String, String> where
the first string is device name, and second string is device hardware address.
We also needed to override getView() and implement the ViewHolder pattern.

It appears that we should NOT stop advertising before starting a scan.
The remote device that is advertising changes its device hardware address as
seen by others who are scanning if we stop/re-start advertising; and doing
scan while advertising is left started appears to be Ok.
For now, we do defensively re-start advertising because we have limited info
on whether all devices will continue advertising in background after screen off
for power save (Nexus 9 appears to do so); the tested devices (Nexus 4 and 9)
seem to tolerate a start to advertise even when it is already advertising.

To allow same code base to run on Nexus 4/5/7 Lollipop (no Peripheral role, no
advertising capabilities) and Nexus 6/9 Lollipop (has Peripheral role and
advertising capabilities), we don't exit if check fails for
 isMultipleAdvertisementSupported()
And we check for null BluetoothGattServer object.

Refactor to make BluetoothLeScanner and BluetoothAdvertiser to be class fields
instead of local fields so they can be re-used each time we start scan or start
advertise.

Improve comments.