Fix handshake to allow successive connection/disconnection; add TextToSpeech
Refactor
1) For Gatt Server, major fix is to do cancelConnection(device) so as not to wait
for what seems to be 30-60s timeout before the connection is actually cancelled
even though the gatt client calls gatt.disconnect() once connection is
established; so we now can do relatively quick successive connections to "Yoble"
the device found with BLE scanning.
For gatt client code, don't need to call gatt.discoverServices() in our use case
because we are just looking for connection established.

2) Gatt Server is instantiated only if isMultipleAdvertisementSupported()
is supported for Peripheral role and BLE advertising, which means we need to check
for certain null field values; so we allow app to be run on Nexus 4/5/7 with
Lollipop, but only in Central role and for BLE scanning.

TTS
Add Text To Speech code to be able to say Yoble. The 'standard' way to set up seems
to use startActivityForResult to check for presence of language data used by TTS
engine, but that seems to momentarily leave the activity and we see NPE, so we
delay by using a delayed post with a handler. The actual set up is done in the
onActivityResult() callback after verifying we have a successful check for TTS data.

Add some null check to fix NPE seen.
Improve toast message wording.
Improve some comments.