First try to get ripple effect
Using a RippleDrawable object as background for TextView.

One big limitation currently is the TextView needs to be
clickable for the ripple animation to work, which then
makes the cards not swipe-able for some reason; seems that
the events are getting consumed by TextView; check out
Intercept Touch Events in a ViewGroup.